requirements.txt
Django==3.2
shortuuid==1.0.1
django-import-export==2.5.0
Pillow==8.2.0